consolidate test mocks and align file names
